Calculates sDNA Integral in a loop of 6 tools in name_map
Still very buggy but this finally proves the new structure of the components works.  Tinkered with Wrapper_Pyshp read and writing a lot.  Added Pyshp 2.20 to fix write record with a dict with missing fields.  sDNA does not round trip user text fields currently, so we really on the order of shapes being preserved.  TODO: Win32 error on trying to delete old shape file