Auto add params works, and plots to GH arcs, with GH gradient comp in between
Still need to fix disappearing legend.  Auto remove params not working, and auto added params appear at the bottom.  Added some arguments to the class decorator function so could pass ghdoc now too, instead of getting it from sc.doc and assigning ghdoc = sc.doc and hoping nothing changed sc.doc in the interim.

Refined how the init, and runscript tool update methods run (on name update and sDNA update), as not all attributes (NickName and Params) on the component are available in the __init__ compared to post __init__.  Added in new method to auto add in the extra tools (for reading Shp and plotting). There are now two tools lists (self.tools and self.mytools) and the return values of the methods
are assigned direct to these to improve readability (not very OOP though?).  The tool runner is now left alone.  The tool factory wrapper should not be called with sDNA_General as a component input it might create an infinite loop, but that's an edge case bordering on deliberate misuse.