Includes new builder .gh that builds all components!
Warning window still comes up, as we are starting from all Params cleared.  If we can copy go and OK and go from there instead, this won't happen.  But it only affects a dev tool (us).  The newly created components don't seem to sync to the whole module options overridden by the builder yet, but test have shown opts are synced - the solutions in GH are just calculated very quickly.  For now I will change the default values for all the auto inserts to False, just as a quick fix to push out a release for testing.