Made output into a __call__able instance of Output
report and report_value are also now replaced by debug, a callable instance of Debug, that is composed from output
Some bloated cludge is now run when string literals are debugged (searching the whole frame locals using inspect)
but it works well, and we don't need to make sure we don't empty generators in the logging anymore.

It was desirable to be able to import from sDNA_GH_launcher for unit testing (defining two separate functions both called output was always poor).
The same code now actually defines two separate classes, as the launcher code must be copied into a GhPython component.  It is now also imported as
it always lived in the sDNA_GH package anyway.  This avoids both unintentional recursion and circular references in code naturally due to the Grasshopper
environment, and saving the launcher code in the package anyway.  But in general, there must be better ways of avoiding recursion!

As there could be multiple components running sDNA_GH_launcher so the root logger is not in __main__.
The root logger is set up when sDNA_GH.tools is first imported (by the first component to run).  Each sDNA_GH_launcher.output
(run as a script if __name__ == '__main__') - True for uncompiled GhPython components but not for compiled ones)
then gets a child of
this logger, and flushes its cache of logging messages into it (using the same code as for the output logging from the first options
override, moved into the flush method of the Output).  The boot lace launcher is now replaced with the same one from sDNA_launcher.

Callable instances could probably be used to get rid of my strange factories and tool functions with attributes!