Major refactor, perhaps biggest to date!
Broke up tools.py into setup.py, skel.tool_runner, skel.quacks_like, tool_factory, dev_tools and tools, helpers.funcs, gdm_from_GH_Datatree, and put more into options_manager and wrapper_pyshp.  Don't put too much code in __init__ or it will always run, even when bits of the package are imported!  Modules that could be a library have had their logging changed to generic from output and debug.  Changed the call in run_tools to **dict so tools can now have default values.  If they're not in component_inputs, the user must add them manually.  Best to set them to None (rust style) and test if var in [None,[None]] separately, as this also catches an unconnected input Param.

So far:
The basic skeleton component just accepts arbitrary args as Param
s.  This calling signiature is unavoidable.  Need to make RunScript then call a customisable method, then a (in another) final method the return args in a separate method, and make the current stack of gets into a while loop, including over locs = locals().copy as per tools.

The tool runner than runs a list of tools (e.g. functions) with certain easily required extra attributes.

So far the tool factory adds in the cache in tools_dict, as a parameter and works its way through name_map.   Everything is set up in tools_dict, except the sDNA tools.  User now specifies a handler in the factory for not found.  e.g. for us, this simply sets the value to sDNAWrapper.

Got rid of the sDNA_General tool.  A wrapper making a recursive tool factory was always going to be tricksy.  Now we just add in sdnageneral as a special name to test, and switch the variable of the component's nickname to tool_name in the args if it's there.  Simples! (touch wood)

Added on to this is Auto inserting of tools into the list according to certain customisable rules.

Added on to / in parallel to auto tool insert tools is Auto updating the actual component Input and Output params.

The Dev tools name_map validation is now very simple, only relying on a list of tool names and name_map.  It should live with the tool_factory, which is the main thing that uses name_map now that sDNA tool wrappers look after their syntax and options themselves (in init and updated thereafter)

As well as breaking down the 5000 line file, the goal of all this is that the base smart component, tool_runner, tool_factory/name_map, auto_insert_tools, and auto_update_Params, and build tool are modular, much more testable, and would also make a nice skeleton GH plug-in for others.