Made most of the SmartComponent into a decorator for RunScript
Used inspect module to get the Argsep of the input function!
args no longer needed to duplicate function args. - TODO: add to tools.
The only way to automatically get the return signature is to grab the source code of a function, strip out the function header, compile it to a code block, set a locals dictionary, and run the code block in exec.  This would support multiple exit points automatically too, but it's not really in the spirit of a simple function call and outside the project scope.  We'll just support returning the values of named local variables in locals for now, with some necessary boiler plate.
Not sure whether to put NickName into SmartComponent in there or not - it's on the component already, so what does it add?