Finalised the smart component Gubbinz
Made a decorator for Custom Inputs as it might be desirable to
customise the call signature, and we can automatically store the old runscript as main as we need to call it.  Custom Outputs currently
need the method adding in manually, but this also needs to be called manually at each exist point, a fact which another decorator might hide.  If imported directly as a base class, the SmartComponent itself is a simple ABC requiring a method called main.