Renamed setup.py to main.py
To avoid confusion with dist utils setup scripts.  main.py is still similar to __main__.py, but is obviously different - main.py is an entry point for Grasshopper and launcher.py, it's not defining a CLI for the package.
Also made dev_tools build from the new default_name_map, to avoid making new components for every nick_name in name_map.