Fixed weird extra keys bug
The bug was occurring during the first options override during the initial import, as metas was at the root level, not one level down at {'metas' : {...}).  This caused the data dictionaries to be treated the same as nested sub dicts at the root level (tool nick names), not data fields of a data node as they should've been.