Fixed sDNA loading bug
Moved windows path to the top of the search order
import_sDNA removes a subfolder called bin at lowest level.
options override now allows sets lists and tuples to be overridden, if the override is an instance of the class of one of the old option's elements
Tested the override to a fake sDNA folder and using the windows path to find the normal one - both work :)
Adjusted launcher.load_modules error checking.