support lists of shapes in shapefiles
 gdms may be a list of gdms, for sDNA Network Radii
A record with a list of shapes in a shape file, now becomes its own gdm (every sub shape with tnhe same dict/value as the record).  Other records and singleton shapes becoming gdms in the list as normal, either side.
Tree helpers should hopefully process this list of dicts in to a Data Tree for Geom.
Needless to say, this needs testing.