Fixes #98 (flattening).
Brought most functions into the TmpShapeFile ... class, and swapped in the patched code in the now method that reads points from shapes on the class.

A sub class init calls generator, which has self.manglers as a default arg val, so the manglers need to be setup before the super call, but the super call is needed to fix the reader from which shapeType is read.  The circular ref in init was closedby using a dangling reference in a different method's body to the hitherto undetermined 2D/3D function that is swapped in by the shapeType value.